update widget according to refactored panel
